football-predict/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.0
scikit-learn==1.5.1
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
fuzzywuzzy==0.18.0
python-Levenshtein-wheels==0.13.2
requests==2.31.0
aiohttp==3.9.5
apscheduler==3.10.1
python-multipart==0.0.9
Jinja2>=3.0.0
pendulum>=3.0.0
cachetools>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic-settings
scrapy==2.5.1
beautifulsoup4==4.10.0
